move utilities, fix nits
